+++ api/src/adapters/aws/__init__.py
@@ -1,3 +1,4 @@
+from .aws_session import get_boto_session
 from .s3_adapter import S3Config, get_s3_client
 
-__all__ = ["get_s3_client", "S3Config"]
+__all__ = ["get_s3_client", "S3Config", "get_boto_session"]
