+++ api/src/api/opportunities_v1/opportunity_schemas.py
@@ -320,7 +320,7 @@ class OpportunityV1Schema(Schema):
 class OpportunityAttachmentV1Schema(Schema):
     download_path = fields.String(
         metadata={
-            "description": "The pre-signed s3 URL to download the attachment",
+            "description": "The URL to download the attachment",
             "example": "https://...",
         }
     )
